[상품등록 API] API 테스트로 전환 (REST-ASSURED 사용)
