+++ src/main/java/com/nahwasa/study/inflearnproductorderservice/product/ProductService.java
@@ -1,9 +1,14 @@
 package com.nahwasa.study.inflearnproductorderservice.product;
 
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Component
+@RestController
+@RequestMapping("/products")
 class ProductService {
     private final ProductPort productPort;
 
@@ -11,10 +16,12 @@ class ProductService {
         this.productPort = productPort;
     }
 
-    @Transactional
-    public void addProduct(AddProductRequest request) {
+    @PostMapping
+    public ResponseEntity<Void> addProduct(@RequestBody AddProductRequest request) {
         final Product product = new Product(request.name(), request.price(), request.discountPolicy());
 
         productPort.save(product);
+
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 }
